adding role services 1.3

+++ PayhouseDragonFly.API/Controllers/User/UserController.cs
@@ -1,13 +1,13 @@
-﻿using Microsoft.AspNetCore.Mvc;
-using PayhouseDragonFly.CORE.ConnectorClasses.Response.authresponse;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
 using PayhouseDragonFly.CORE.ConnectorClasses.Response;
+using PayhouseDragonFly.CORE.ConnectorClasses.Response.authresponse;
+using PayhouseDragonFly.CORE.ConnectorClasses.Response.BseResponse;
 using PayhouseDragonFly.CORE.DTOs.loginvms;
 using PayhouseDragonFly.CORE.DTOs.RegisterVms;
+using PayhouseDragonFly.CORE.DTOs.userStatusvm;
+using PayhouseDragonFly.INFRASTRUCTURE.Services.ExtraServices.RoleChecker;
 using PayhouseDragonFly.INFRASTRUCTURE.Services.IServiceCoreInterfaces.IUserServices;
-
-using PayhouseDragonFly.CORE.ConnectorClasses.Response.BseResponse;
-using Microsoft.AspNetCore.Authorization;
-using PayhouseDragonFly.CORE.DTOs.Roles;
 
 namespace PayhouseDragonFly.API.Controllers.User
 {
@@ -17,10 +17,12 @@ namespace PayhouseDragonFly.API.Controllers.User
 
     public class UserController : ControllerBase
     {
-        private readonly IUserServices _userServices;
-        public UserController(IUserServices userServices)
+        public readonly IUserServices _userServices;
+        private readonly IRoleChecker _rolechecker;
+        public UserController(IUserServices userServices, IRoleChecker rolechecker)
         {
             _userServices = userServices;
+            _rolechecker = rolechecker;
         }
 
 
@@ -30,6 +32,7 @@ namespace PayhouseDragonFly.API.Controllers.User
         public async Task<authenticationResponses> Authenticate(loginvm loggedinuser)
         {
 
+            
             return await _userServices.Authenticate(loggedinuser);
 
         }
@@ -47,8 +50,16 @@ namespace PayhouseDragonFly.API.Controllers.User
 
         public async Task<BaseResponse> GetAllUsers()
         {
+            var roleidretured = _rolechecker.Returnedrole().Result;
+            if (roleidretured == 1)
+            {
+                return await _userServices.GetAllUsers();
+            }
+            else
+            {
 
-            return await _userServices.GetAllUsers();
+                return new BaseResponse("120", "You have no permission access this", null); 
+            }
         }
 
 
@@ -57,7 +68,16 @@ namespace PayhouseDragonFly.API.Controllers.User
         [Route("DeleteUsers")]
         public async Task<BaseResponse> DeleteUser(string usermail)
         {
-            return await _userServices.DeleteUser(usermail);
+            var roleidretured = _rolechecker.Returnedrole().Result;
+            if (roleidretured == 1)
+            {
+                return await _userServices.DeleteUser(usermail);
+            }
+            else
+            {
+
+                return new BaseResponse("120", "You have no permission access this", null);
+            }
         }
 
 
@@ -69,7 +89,16 @@ namespace PayhouseDragonFly.API.Controllers.User
         public async Task<BaseResponse> ConfirmUserAccount(string useremail)
         {
 
-            return await _userServices.ConfirmUserAccount(useremail);
+            var roleidretured = _rolechecker.Returnedrole().Result;
+            if (roleidretured == 1)
+            {
+                return await _userServices.ConfirmUserAccount(useremail);
+            }
+            else
+            {
+
+                return new BaseResponse("120", "You have no permission access this", null);
+            }
         }
 
         [Authorize(AuthenticationSchemes = "Bearer")]
@@ -102,6 +131,7 @@ namespace PayhouseDragonFly.API.Controllers.User
         [Route("Add_Department")]
         public async Task<BaseResponse> AddDepartment(AddDepartmentvms addDepartmentvms)
         {
+
             return await _userServices.AddDepartment(addDepartmentvms);
 
         }
@@ -111,9 +141,14 @@ namespace PayhouseDragonFly.API.Controllers.User
 
         public async Task<BaseResponse> GetAllDepartment()
         {
+            var roleidretured = _rolechecker.Returnedrole().Result;
+            if (roleidretured == 1)
+            {
 
-            return await _userServices.GetAllDepartment();
+                return await _userServices.GetAllDepartment();
         }
+            return new BaseResponse("120", "You have no permission access this", null);
+    }
 
 
         [HttpGet]
@@ -124,9 +159,69 @@ namespace PayhouseDragonFly.API.Controllers.User
             return await _userServices.TestMail(testmail);
         }
 
+        [HttpPost]
+        [Route("Getdepartmentbyid")]
+        public async Task<BaseResponse> GetDepartmentByID(int departmentid)
+        {
+            return await _userServices.GetDepartmentByID(departmentid);
+        }
+        [Authorize(AuthenticationSchemes = "Bearer")]
+        [HttpPost]
+        [Route("Suspend_user")]
+        public async Task<BaseResponse> SuspendUser(suspendUservm vm)
+
+        {
+            return await _userServices.SuspendUser(vm);
+
+        }
+        [Authorize(AuthenticationSchemes = "Bearer")]
+        [HttpPost]
+        [Route("Change_UserStatus")]
+       public async Task<BaseResponse> ChangeUserStatus(userStatusvm vm)
+        {
+            var roleidretured = _rolechecker.Returnedrole().Result;
+            if (roleidretured == 1)
+            {
+                return await _userServices.ChangeUserStatus(vm);
+            }
+            return new BaseResponse("120", "You have no permission access this", null);
+        }
+        [Authorize(AuthenticationSchemes = "Bearer")]
+        [HttpPost]
+        [Route("GetUserActiveStatusByid")]
+        public async Task<BaseResponse> GetUserActiveStatusByid(string userid)
+        {
+            return await _userServices.GetUserActiveStatusByid(userid);
+        }
 
 
+        [Authorize(AuthenticationSchemes = "Bearer")]
+        [HttpGet]
+        [Route("LoggedInUser")]
+        public async Task<BaseResponse> GetLoggedInUser()
+        {
+            return await _userServices.GetLoggedInUser();
+        }
 
+
+        [Authorize(AuthenticationSchemes = "Bearer")]
+        [HttpPost]
+        [Route("EditUser")]
+        public async Task<BaseResponse> EditUserDetails(RegisterVms edituservm, string userid)
+        {
+
+            return await _userServices.EditUserDetails(edituservm, userid);
+        }
+
+
+        [Authorize(AuthenticationSchemes = "Bearer")]
+        [HttpPost]
+        [Route("ActivateUserAccount")]
+        public async Task<BaseResponse> ActivateUser(string useremail)
+        {
+            return await _userServices.ActivateUser(useremail);
+
+        }
 
     }
